+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Journal des modifications
 =========================
 
 ## [X.X] en cours
+- Prise en compte du déplacement de Pm_TolStQ du CCM vers le OPTI pour les initialisations de type Saint-Venant (g1.3)
 - Amélioration performances lecture binaire en utilisant `np.frombuffer` (au lieu de struct)
 - Correction plantage lecture RCAL si aucun résultat aux branches